refactor: Implement 6 contract feedback fixes and add integration tests for FastAPI churn & vehicle prediction

+++ SpringBoot/app/src/main/java/com/car/app/ai/AiClient.java
@@ -245,7 +245,7 @@ public class AiClient {
         @com.fasterxml.jackson.annotation.JsonProperty("predicted_condition")
         private Double predictedCondition;
         @com.fasterxml.jackson.annotation.JsonProperty("predicted_mmr")
-        private Long predictedMmr;
+        private Double predictedMmr;
     }
 
     @Getter
